updated files for channel names
channel-order contains the all the nodes used by the initial recording but with modified names to follow the naming convention of Standard 10_20 Montage in the MNE lib. channel-order(viz) is the same as channel order but with row 58 and 62 deleted due to nodes CB1 and CB2 not being in the the naming convention of Standard 10_20 Montage in the MNE lib.